pyproject.toml
[project]
name = "mcp-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dateparser>=1.2.2",
    "groq>=0.37.1",
    "langchain>=1.2.0",
    "langchain-groq>=1.1.1",
    "mcp>=1.25.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "supabase>=2.27.0",
]
